requirements.txt
pandas
openpyxl
dotenv
google-genai